Combine experiment_manager.py and agent_experiment_manager.py into a single file
